+++ crates/emmylua_parser/src/syntax/node/doc/tag.rs
@@ -1,13 +1,14 @@
 use crate::{
-    kind::LuaSyntaxKind, syntax::traits::LuaAstNode, BinaryOperator, LuaAstChildren, LuaAstToken,
-    LuaAstTokenChildren, LuaBinaryOpToken, LuaDocVersionNumberToken, LuaDocVisibilityToken,
-    LuaGeneralToken, LuaKind, LuaNameToken, LuaNumberToken, LuaPathToken, LuaStringToken,
-    LuaSyntaxNode, LuaTokenKind, LuaVersionCondition,
+    kind::LuaSyntaxKind,
+    syntax::{traits::LuaAstNode, LuaDocDescriptionOwner},
+    BinaryOperator, LuaAstChildren, LuaAstToken, LuaAstTokenChildren, LuaBinaryOpToken,
+    LuaDocVersionNumberToken, LuaDocVisibilityToken, LuaGeneralToken, LuaKind, LuaNameToken,
+    LuaNumberToken, LuaPathToken, LuaStringToken, LuaSyntaxNode, LuaTokenKind, LuaVersionCondition,
 };
 
 use super::{
-    description::{LuaDocDescriptionOwner, LuaDocDetailOwner},
-    LuaDocAttribute, LuaDocGenericDeclList, LuaDocOpType, LuaDocType, LuaDocTypeList,
+    description::LuaDocDetailOwner, LuaDocAttribute, LuaDocGenericDeclList, LuaDocOpType,
+    LuaDocType, LuaDocTypeList,
 };
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]